create expense and income
